requirements.txt
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
backports.zoneinfo==0.2.1
certifi==2023.11.17
cffi==1.16.0
channels==3.0.5
channels-redis==4.1.0
constantly==23.10.4
cryptography==41.0.7
daphne==3.0.2
distro==1.8.0
Django==4.2.8
django-filter==23.4
django-rest-knox==4.2.0
djangorestframework==3.14.0
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
msgpack==1.0.7
openai==1.3.7
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
pyOpenSSL==23.3.0
python-decouple==3.8
pytz==2023.3.post1
redis==5.0.1
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
tqdm==4.66.1
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.8.0
zope.interface==6.1
